# fastserver/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for a Go web mini project.",
        "title": "Go Web fastgin API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "192.168.123.214:8088",
    "basePath": "/api/",
    "paths": {
...
        "/api/batch_delete": {
            "delete": {
                "description": "Delete multiple APIs by their IDs",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "API"
                ],
                "summary": "Batch delete APIs",
                "parameters": [
                    {
                        "description": "Delete API request",
                        "name": "apiIds",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/bean.DeleteApiRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseBody"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseBody"
                        }
                    }
                }
            }
        },
        "/api/tree": {
            "get": {
                "description": "Get the API tree categorized by the Category field",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "API"
                ],
                "summary": "Get API tree",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseBody"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseBody"
                        }
                    }
                }
            }
        },
...
        "/apis": {
            "get": {
                "description": "Get a list of APIs",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "API"
                ],
                "summary": "Get API list",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Request method",
                        "name": "method",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Request path",
                        "name": "path",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Category",
                        "name": "category",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Creator",
                        "name": "creator",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "pageNum",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseBody"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseBody"
                        }
                    }
                }
            }
        },
...
        "/operation_logs": {
            "get": {
                "description": "Get a list of operation logs",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OperationLog"
                ],
                "summary": "Get operation log list",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Request method",
                        "name": "method",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Request path",
                        "name": "path",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Category",
                        "name": "category",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Creator",
                        "name": "creator",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "pageNum",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseBody"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseBody"
                        }
                    }
                }
            }
        },
        "/operation_logs/batch_delete": {
            "delete": {
                "description": "Delete multiple operation logs by their IDs",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OperationLog"
                ],
                "summary": "Batch delete operation logs",
                "parameters": [
                    {
                        "description": "Delete operation log request",
                        "name": "operationLogIds",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/bean.DeleteOperationLogRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseBody"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseBody"
                        }
                    }
                }
            }
        },
...
        "/role/{roleId}/apis": {
            "get": {
                "description": "Get the APIs for a role by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Role"
                ],
                "summary": "Get role APIs by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Role ID",
                        "name": "roleId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseBody"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseBody"
                        }
                    }
                }
            },
            "put": {
                "description": "Update the APIs for a role by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Role"
                ],
                "summary": "Update role APIs by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Role ID",
                        "name": "roleId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update role APIs request",
                        "name": "apis",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/bean.UpdateRoleApisRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseBody"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseBody"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "bean.DeleteApiRequest": {
            "type": "object",
            "properties": {
                "apiIds": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
        "bean.DeleteOperationLogRequest": {
            "type": "object",
            "properties": {
                "operationLogIds": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
        "bean.UpdateRoleApisRequest": {
            "type": "object",
            "properties": {
                "apiIds": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
        "controller.ResponseBody": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}